+++ src/Sulu/Bundle/ContentBundle/Repository/ResourceLocatorRepository.php
@@ -11,10 +11,14 @@
 
 namespace Sulu\Bundle\ContentBundle\Repository;
 
+use Sulu\Bundle\DocumentManagerBundle\Bridge\DocumentInspector;
 use Sulu\Component\Content\Compat\StructureInterface;
 use Sulu\Component\Content\Compat\StructureManagerInterface;
+use Sulu\Component\Content\Document\LocalizationState;
 use Sulu\Component\Content\Types\Rlp\ResourceLocatorInformation;
 use Sulu\Component\Content\Types\Rlp\Strategy\RlpStrategyInterface;
+use Sulu\Component\DocumentManager\Behavior\Mapping\ParentBehavior;
+use Sulu\Component\DocumentManager\DocumentManagerInterface;
 
 /**
  * resource locator repository.
@@ -32,6 +36,16 @@ class ResourceLocatorRepository implements ResourceLocatorRepositoryInterface
     private $rlpStrategy;
 
     /**
+     * @var DocumentManagerInterface
+     */
+    private $documentManager;
+
+    /**
+     * @var DocumentInspector
+     */
+    private $documentInspector;
+
+    /**
      * @var string[]
      */
     private $apiBasePath = [
@@ -45,10 +59,14 @@ class ResourceLocatorRepository implements ResourceLocatorRepositoryInterface
      */
     public function __construct(
         RlpStrategyInterface $rlpStrategy,
-        StructureManagerInterface $structureManager
+        StructureManagerInterface $structureManager,
+        DocumentManagerInterface $documentManager,
+        DocumentInspector $documentInspector
     ) {
         $this->rlpStrategy = $rlpStrategy;
         $this->structureManager = $structureManager;
+        $this->documentManager = $documentManager;
+        $this->documentInspector = $documentInspector;
     }
 
     /**
@@ -61,7 +79,8 @@ class ResourceLocatorRepository implements ResourceLocatorRepositoryInterface
         $title = $this->implodeRlpParts($structure, $parts);
 
         if ($parentUuid !== null) {
-            $parentPath = $this->rlpStrategy->loadByContentUuid($parentUuid, $webspaceKey, $languageCode, $segmentKey);
+            $parentDocument = $this->documentManager->find($parentUuid, $languageCode);
+            $parentPath = $this->rlpStrategy->loadByContent($this->getNonGhostAncestor($parentDocument));
             $result = $this->rlpStrategy->generate($title, $parentPath, $webspaceKey, $languageCode, $segmentKey);
         } elseif ($uuid !== null) {
             $result = $this->rlpStrategy->generateForUuid($title, $uuid, $webspaceKey, $languageCode, $segmentKey);
@@ -154,5 +173,24 @@ class ResourceLocatorRepository implements ResourceLocatorRepositoryInterface
         $title = substr($title, 0, -1);
 
         return $title;
+    }
+
+    /**
+     * Returns the first ancestor-or-self of the given document which is not a ghost-document.
+     * If all ancestor documents are ghost-documents, the root document is returned.
+     *
+     * @param object    $document
+     *
+     * @return object
+     */
+    private function getNonGhostAncestor($document)
+    {
+        while ($this->documentInspector->getLocalizationState($document) === LocalizationState::GHOST
+            && $document instanceof ParentBehavior
+        ) {
+            $document = $document->getParent();
+        }
+
+        return $document;
     }
 }